Refactor: Initial commit
This commit introduces the initial structure of the shapes calculator application. It includes the basic setup for calculating properties of triangles, rectangles, squares, and circles.

Key changes:
- Adds activities and layouts for each shape (Triangle, Rectangle, Square, Circle).
- Implements a main menu (`MenufigurasActivity`) to navigate between the different shape calculators.
- Moves shape-specific logic and views from a single `MainActivity` into their respective dedicated activities and layouts.
- Updates the Android Manifest to declare the new activities and sets the menu as the launcher activity.
- Renames layout files and classes for better organization and clarity.

+++ app/src/main/java/com/example/triangulo4b/Contrato/ContratoCirculo.kt
@@ -3,13 +3,13 @@ package com.example.triangulo4b.Contrato
 interface ContratoCirculo {
 
     interface Modelo{
-        fun areaCirculo(radio: Float): Double
-        fun perimetroCirculo(radio: Float): Double
+        fun areaCirculo(radio: Float): Float
+        fun perimetroCirculo(radio: Float): Float
         fun validaCirculo(radio: Float): Boolean
     }
     interface Vista{
-        fun showAreaCirculo(area: Double)
-        fun showPerimetroCirculo(perimetro: Double)
+        fun showAreaCirculo(area: Float)
+        fun showPerimetroCirculo(perimetro: Float)
         fun showErrorCirculo(msg: String)
     }
     interface Presentador{
